+++ studit/src/main/java/com/example/studit/config/exception/BaseException.java
@@ -1,4 +1,4 @@
-package com.example.studit.config.swagger;
+package com.example.studit.config.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class BaseException {
+public class BaseException extends Exception{
     private BaseResponseStatus status;
 }
